RX2 Implementation

+++ dbflow/src/main/java/com/raizlabs/android/dbflow/list/FlowCursorList.java
@@ -103,7 +103,7 @@ public class FlowCursorList<TModel> implements
     }
 
     @Override
-    public FlowCursorIterator<TModel> iterator(int startingLocation, int limit) {
+    public FlowCursorIterator<TModel> iterator(int startingLocation, long limit) {
         return new FlowCursorIterator<>(this, startingLocation, limit);
     }
