+++ dbflow/src/main/java/com/raizlabs/android/dbflow/list/FlowQueryList.java
@@ -373,7 +373,7 @@ public class FlowQueryList<TModel> extends FlowContentObserver
     }
 
     @Override
-    public FlowCursorIterator<TModel> iterator(int startingLocation, int limit) {
+    public FlowCursorIterator<TModel> iterator(int startingLocation, long limit) {
         return new FlowCursorIterator<>(this, startingLocation, limit);
     }
